requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==4.2
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==24.2
djangorestframework==3.15.1
docx2pdf==0.1.8
drf-yasg==1.21.7
fonttools==4.53.0
fpdf==1.7.2
fpdf2==2.7.9
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
lxml==5.2.2
packaging==24.0
pdfkit==1.0.0
pillow==10.3.0
psycopg2-binary==2.9.9
Pygments==2.18.0
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
